inf_tech/requirements.txt
Flask
viberbot
filestack-python
requests